sprint 8. 18.06 работа над новым проектом-8. финал
